// src/app/model/descriphistoria.ts
export class DescripHistoria{
    cdDescripHistoria: number;
    dsVivienda: string;
    nmMascota: number;
    fechaDes: string;
    fechaVacu: string;
    dsAlimento: string;
    dsFrecuencia: string;
    dsReproductivo: string;
    dsTratamientos: string;
    dsEnfermedades: string;
    dsCirugias: string;
    dsMotivo: string;
    dsTemperatura: string;
    dsTemperamento: string;
    dsPeso: string;
    dsFc: string;
    dsFr: string;
    dsTllc: string;
    dsMm: string;
    dsPulso: string;
    dsActitud: string;
    dsHidratacion: string;
    dsNervioso: string;
    dsPiel: string;
    dsMuscu: string;
    dsRespira: string;
    dsCardio: string;
    dsDiges: string;
    dsUri: string;
    dsLinfoi: string;
    dsReprod: string;
    dsOjos: string;
    dsOidos: string;
    dsHallaz: string;
    dsConsulta: string;
    dsTerapeu: string;
    dsSegu: string;
    cdHistoria: number;
}